requirements.txt
django==2.1.2
markdown==3.2.2
beautifulsoup4==4.9.2
requests==2.24.0
djangorestframework==3.11.1
bs4==0.0.1
djangoextensions==3.0.9
html5lib==1.1
django-filter==2.3.0
Pygments==2.7.1
httpie==2.2.0
lxml==4.5.2
